requirements.txt
python-telegram-bot==13.15
requests==2.31.0
reportlab==4.2.0
openpyxl==3.1.2
flask==3.0.0
